refactor: Remove TodosService from additional-services and create a dedicated todo-list-service
